Split: take away profiling from __init__.py
